[14.0][IMP] account_wallet_coupon: Improve behaviour
- Adds a check if no wallet with filled in coupon code is found
- Initialize payment amount with wallet balance if invoice amount is >
