+++ account_wallet_coupon/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import common
 from . import test_wallet_coupon
+from . import test_wallet_coupon_payment
